feat: run CI on Windows
We'll see how performant this is. In the old repo, in Travis, Windows
was dog slow. Let's see if it is fast enough as-is; if it isn't there
are a number of Pro-Tips™ linked to in the related issues.

Seeing as I am splitting out the Windows test run (and on Windows we are
only going to run tests, not format-checking and linting), I also split
out the rest. We'll see if this improves or worsens the speed of
feedback; we're running stuff in parallel (good), but each parallel job
duplicates the work of checking out, running `yarn install` etc.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/37