test: attempt fix for test failure on Windows
Flying a bit blind here but we have one failure:

```
FAIL projects/npm-tools/packages/npm-scripts/test/utils/filterChangedFiles.js
  ● filterChangedFiles() › when LIFERAY_NPM_SCRIPTS_WORKING_BRANCH_NAME is set › returns only changed files, unless it detects a @liferay/npm-scripts update

    expect(received).toEqual(expected) // deep equality

    - Expected  - 3
    + Received  + 1

    - Array [
    -   "code",
    - ]
    + Array []

      at Object.<anonymous> (test/utils/filterChangedFiles.js:178:38)
```

That is occurring in the Windows environment on GitHub actions:

https://github.com/liferay/liferay-frontend-projects/pull/85/checks?check_run_id=1166887150

But passed fine in the Windows enviroment on Travis:

https://travis-ci.org/github/liferay/liferay-npm-tools/jobs/730211811

So, going to take a stab in the dark and see if using NUL as a path
separator fixes things. I am only 20% optimistic that it might.  The
reason I am even looking at this line as potentially implicated is that
'\n' might not be the right separator (ie. it could be '\r\n').