test: attempt fix for several Windows failures
On the GitHub Actions Windows environments, there is a directory at
`C:\modules`, and our code would happily walk up and find it. This
reveals a subtle bug in the old implementation, because we don't really
want to find _siblings_ that are called "modules" as we walk up the tree;
we want to want walk up and return when _the current node_ is called
"modules".

So this commit implements that change. I am hoping this will fix the
issue on Windows.