feat: add support for formatting and linting TypeScript
Note that the goal here is to format and lint TypeScript in
liferay-portal, but we're "dog-fooding" in this monorepo already, even
though there isn't any TypeScript in here (yet... when we bring across
the js-toolkit packages TypeScript will start coming in here too). Note
how we can lint JS just fine with the TS parser.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/96