refactor: try applying TS-specific parser only to ".ts" and ".tsx"
This is an attempt to address some of the problems described in this
comment:

https://github.com/liferay/liferay-frontend-projects/pull/97#issuecomment-701517441

Namely, that some of our JS files have namespaced JSX attributes in
them (eg. `<use xlink:href="..." />`), which TypeScript doesn't support
at this time (and might not in the future).

Test plan: I'll test this in liferay-portal, but for local testing, I
added a sample TS file and verified that we see formatting/linting
errors in there (when present). Lint suppressions also work.

We can remove this file in the future, after we have some actually
real/used TS in this repo.