docs(scripts): over-promise and almost-deliver
As suggested here:

https://github.com/liferay/liferay-frontend-projects/pull/100#discussion_r498340484

We eventually want to use liferay-npm-scripst in more places, although
at the moment we are only using it in liferay-portal and liferay-learn
(and some brave souls have experimented with using it inside
workspaces), so we may as well change the language to reflect the
direction we're heading in. One less thing to update in the future then.