fix(npm-tools): make Prettier overrides work with TS
So, I thought I had this working in:

    https://github.com/liferay/liferay-frontend-projects/pull/101

(see the test plan), but when I got this into liferay-portal in:

    https://github.com/liferay-frontend/liferay-portal/pull/415

I found it wasn't working (the `else` formatting wasn't applying). So
I made the edits in this commit directly in the node_modules folder to
realllllly see it working.

First step: switch in the TS parser when dealing with a TS file.

Second step: observe the same problem we saw (and fixed) in
d8b03088be5c76fed — namely, that the TS parser doesn't put `start` and
`end` properties on its tokens. We can switch to `range[0]` and
`range[1]` respectively, making the rule work with both parsers.