docs: encourage people to dog-food local packages
A few places where we're telling people to use the last-published
version of the changelog-generator (via `npx`) tell them to use `yarn
run` instead; this means they'll get the current version in the workspace
instead.

About the only time this won't work is if you haven't run `yarn` yet,
which I'm not particularly worried about, or you have broken the local
copy of the generator (and that's one of the main reasons why we like
dog-fooding anyway; to find out when we've broken things).