style: fix formatting on imported projects
This shows one of the nice things about the monorepo: even though we
have two "products" in here (the 2.x series living under
"maintenance/projects/" and the 3.x series living under "projects/"), we
can apply uniform formatting to both of them that matches the formatting
used everywhere else in the monorepo. And this is true even though 2.x
is isolated from top-level Yarn workspaces.