docs: applies review feedback
Avoid noise on "inline js on the JSP doc"
English wordsmith
Evergreen links
Adds warnings about the AUI/YUI js/taglib usage
`JSP context` -> `JSP files`
`modern JS` -> `ES module` for compliance with the glossary
adds the aui:script require output
Adds some output examples
remove commas from headings and improve the explanation of the benefits of using `aui:script`
replace `es6` occurences with `modern JS` for not reproduce the idea that we are following the old ES2015 spec
'we' -> 'We'
strengthen the resulted variable of a imported resource as a normalized and camel-case name when using require/amd-loaded feature on aui:script
Adds an use case for `Inline JS on the JSP`
Avoids `Pure JS on JSP` term and use `inline JS on JSP`
Avoid using `taglib` when reffering to just one tag. `taglib` is a library of tags. It also fixes a english typo
Fix sandbox section

* Omits jquery and underscore mention since We don't recommend using it anymore
* Adds an example of sandbox true reinforcing the IIFE behavior