chore(js-toolkit): make many adjustments to get lints running
Because the version of the `@typescript-eslint/eslint-plugin` package
has changed, we had lots of "missing rule" errors due to breaking
changes as mentioned in a number of places:

- https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/camelcase.md
- https://github.com/typescript-eslint/typescript-eslint/issues/2054
- https://github.com/typescript-eslint/typescript-eslint/issues/1854

This was producing approximately one zillion errors (on basically every
".ts" file) of the form:

    1:1  error  Definition for rule '@typescript-eslint/ban-ts-ignore' was not found
    1:1  error  Definition for rule '@typescript-eslint/camelcase' was not found
    1:1  error  Definition for rule '@typescript-eslint/class-name-casing' was not found
    1:1  error  Definition for rule '@typescript-eslint/interface-name-prefix' was not found

The first rule got renamed to "ban-ts-comments" and the others got
deleted and replaced by "naming-convention".

So, I updated those rules, removed the `root` directives which should
not apply, added missing "notice/notice" config, and deleted things
which should be inherited from above. With this, lints run without
spurious errors, as far as I can tell.

In the next few commits I'll strip out the copyright headers which were
stale (due to the missing "notice/notice" config), run autofixes, and
finish up with manual fixes.