chore(js-toolkit): avoid ESLint vs Prettier conflict
As described in the linked issue, this is a temporary workaround:

    https://github.com/liferay/liferay-frontend-projects/issues/128

Prettier wants:

```ts
export interface X {
  /**
   * Yo!
   */
  thing: string;
}
```

while ESLint wants:

```ts
export interface X {

  /**
   * Yo!
   */
  thing: string;
}
```

(There's a similar fight for `//` comments as well.)

Now, js-toolkit has a couple dozen of these darn comments, so I
either add a suppression or I delete the comments. I'm going to add a
suppression, and create this issue to remind me to remove it.

The definitive solution will be to dog-food liferay-npm-scripts
and use it in the monorepo to do the formatting. As seen in
bf5d84c475fb478bfa3378b9ab54d012e935a063, we don't have this problem in
liferay-portal thanks to liferay-npm-scripts (because it runs ESLint
+ Prettier + post-processing, we can resolve little conflicts like
this). Obviously, there's a benefit to dog-fooding anyway, especially
now that other projects like liferay-learn are using npm-scripts too
(ie. it is useful to us to be dog-fooding it all the time, so that we
find out if we ever break the non-liferay-portal use case).