refactor: adjust build to work in CI
So, unlike the other projects imported previously, you can't run the
tests on the js-toolkit projects without doing a build first.

So, we're going to run "build" for _all_ projects via "yarn workspaces
run build". This means that we have to add no-op stubs in the projects
that don't have them.

Note: this isn't ideal and will probably be refactored very soon,
because js-toolkit has a top-level "build" which builds everything, and
then per-package "builds" that build alone. So, if we visit every
workspace, we're actually building js-toolkit twice. This is ok, for
now, only because it is a small project with 5 packages and so builds
fast.