test(js-toolkit): make v2.x tests pass in CI
Example failure here[0].

As explained in the comment, the behavior of this code is sensitive to
the environment that you run in, so I'm making the test looser to work
both locally and in CI.

In the monorepo, the `getRenderer()`[1] method is going to try to get
"node-sass" first and it is going to find it because it is upstairs in
the top-level "node_modules" folder.

In CI, when we run the v2.x tests, we only do a yarn install inside
"maintenance/projects/js-toolkit", because it doesn't participate in the
top-level Yarn workspaces. So in CI, `getRenderer()` is going to try to
get "node-sass", fail, and fallback to "sass", which is bundled more
locally.

As such, we need to loosen our assertion about what the renderer
description looks like, as well as make the assertion about the expected
output whitespace-agnostic, seeing as "node-sass" and "sass" make
slightly different choices about whitespace. I figure this looseness is
more future-proof and flexible than the alternatives, because it is more
likely to keep the suite green even if we refactor and restructure the
relationship between packages in the monorepo; and, at the end of the
day, the `getRenderer()` is fundamentally dynamic in nature.

[0]: https://github.com/liferay/liferay-frontend-projects/pull/133/checks?check_run_id=1220537713
[1]: https://github.com/liferay/liferay-frontend-projects/blob/47c938cc1c1a7a2837453eb96535fcec9a6492a3/maintenance/projects/js-toolkit/packages/liferay-npm-bundler-loader-sass-loader/src/index.js#L31-L58