test(js-toolkit): fix failing test on Windows
As seen here:

https://github.com/liferay/liferay-frontend-projects/pull/133/checks?check_run_id=1221124885

```
FAIL projects/js-toolkit/packages/liferay-js-toolkit-core/src/file/__tests__/FilePath.test.ts
  ● in posix systems › toDotRelative works

    expect(received).toEqual(expected) // deep equality

    Expected: "./a/path"
    Received: "./a\\path"

      41 |      it('toDotRelative works', () => {
      42 |              expect(new FilePath('').toDotRelative().asNative).toEqual('.');
    > 43 |              expect(new FilePath('a/path').toDotRelative().asNative).toEqual(
         |                                                                      ^
      44 |                      './a/path'
      45 |              );
      46 |

      at Object.<anonymous> (packages/liferay-js-toolkit-core/src/file/__tests__/FilePath.test.ts:43:59)
```

So, prior to this commit, we had tests that tried to exercise both POSIX
and Windows environments by faking the environment. ie. you could be on
your Linux machine and run the "Windows" examples and see them pass, but
you obviously weren't testing real Windows and we were just setting
`nativeIsPosix` to `false` (and vice a versa for the POSIX examples).

Well, now that we run in a real Windows environment we see the above
failure and it is because even though we set `nativeIsPosix` to `true`,
it's not really true. We hit a `path.normalize()` call in the middle of
that function that turns our POSIX separators back to Windows
separators, as per the docs:

> they are replaced by a single instance of the platform-specific
> path segment separator ... Since Windows recognizes multiple path
> separators, both separators will be replaced by instances of the
> Windows preferred separator

https://nodejs.org/api/path.html#path_path_normalize_path

Obviously, `path.normalize()` is not affected by out little stub.

So, the solution implemented in this commit is to remove the stubbing
and instead only run the native platform tests. This means that when you
run on Windows, you will run the Windows tests, and when you run on
Linux or macOS, you will run the POSIX tests. And because CI runs both,
we still cover all bases before anything gets merged.