refactor(js-toolkit): remove "chalk" devDependency
I'm going to hoist as many dev dependencies to the top of the monorepo
but I don't want to do that with chalk yet as we already have a huge
number of places that depend on it transitively. The output of `yarn why
chalk` can be seen in this 399-line gist:

https://gist.github.com/wincent/a86a06a9e3eaceafe6d3fff27bc30a76

Rather than adding to the chaos, let's remove "chalk" as a dev
dependency from the js-toolkit because it is barely used at all: only
`scripts/util/report.js` uses it, which is itself only imported by one
other file. Most of it is dead code (because `success()` and `warn()`
are never used), and `abort()` is only used once, in a code path that
should in theory never be executed unless something goes wrong.

So, if we strip out the dead code, that means we're not using `green` or
`yellow` from "chalk", and we may as well not keep it around just for a
`red` that is probably never used. (And in any case, the big-red-cross
emoji and the massive ugly stack trace that `console.error` will print
out is probably enough.)

Speaking of stack traces, we don't want to print out three in a row, so
we shouldn't call `console.error()` three times; I collapsed them into a
single call. Likewise, I got rid of the now-unneeded lint suppression
(`console.log()` is not allowed, but `console.error()` is just fine).