refactor(js-toolkit): hoist mem-fs and mem-fs-editor
Note that we're jumping from v6 to v7 in mem-fs-editor, and the project
unfortunately doesn't have a changelog or release notes published here:

https://github.com/SBoudrias/mem-fs-editor/releases

So, the only way to see if it breaks anything is to update it and see.
Seeing as it's used only in the test suite, I'm not too worried.

Add these two to the list of probably-not-justified dev dependencies
that are used in exactly one place, and which we can look at removing or
replacing in a future audit.