fix(eslint-config): special case import of "."
I don't like these "." imports as there are not very idiomatic; I guess
they're not common because they are subtle and potentially confusing.
If you are in a folder and "./index" is the manifest of everything
exportable in that folder, then that would normally include the file
doing the import, which is circular. So the cognitive load of thinking
that through is higher than it need be.

Anyway, even though I don't like them, it doesn't mean I should fix the
bug. So here goes.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/129