docs(js-toolkit): update links prior to release
As I want to cut a release (just a patch release, to prove that
releasing from the monorepo works, and to update the README and metadata
on the package pages on the npm registry) I'm just sweeping through and
updating links and other information that needs to be tweaked now that
we're in the monorepo.

The one link I didn't update was the link to the wiki, because we
haven't migrated that yet (we still need to decide whether to migrate it, or
whether to move the contents into a `docs/` folder).

For consistency, snuck in one edit to the top-level README to make sure
we use the same alt text on our CI badges everywhere.