style(js-toolkit): run `yarn format`
For some reason, Lerna, I think, changed the formatting of this file
when I ran the release. You can see the unwanted formatting change
sneaking in in d341f385c91a7d50168a6ce8be978d2f03f6bedc.

Will keep an eye on this to see if it happens again in the next release.