docs(js-toolkit): clean up external references to 2.x wiki
Note that I removed some links because there is not much point in
having links from the 3.x packages to the 2.x documentation.

At some point in the future, we have the option of moving all the docs
into the main "projects/js-toolkit/" folder, and calling out the parts
of it that refer to 3.x and the parts which refer to 2.x, but for now,
we just have 2.x documentation.