docs(js-toolkit): convert MediaWiki style links to Markdown links
Done with a couple of complicated Vim stuff...

Something like:

    :Ack \[\[[^\]]+\]\]

to find the files.

Then inside each one:

    :%s#\v\[\[([^\]]+)\]\]#[\1](\1.md)#gc

And then converting spaces to hyphens with:

    :'<,'>s/\%V /-/g

ie. jumping between `:Ack` results with Up/Down, then "f(" to find link
target, "vi(" to select it, and Up + Return to run the above
substitution (sadly, all this doesn't seem to actually work in a macro).