docs(js-toolkit): retarget remaining links to point at monorepo
Like I said I was going to do here:

https://github.com/liferay/liferay-frontend-projects/pull/144#issuecomment-705566543

As far as I know, nothing else major to be tweaked at this point. For
example, there are issue links, which can all stay pointed at the old
repo and will continue to work. There is a schema link that we could
change too, but I don't want to do that as part of a documentation
update.