feat(npm-scripts): teach VSCode wrapper about this repo
So that we can get Liferay-specific formatting in this repo as well.

This is not a particularly "clean" solution, but nothing in this script
is clean...

We detect liferay-portal by looking at the directory structure, but we
detect the monorepo by looking for a "sentinel" dotfile in the root. We
can do this in the monorepo because it is "our" repo, unlike
liferay-portal, where many more people work, and adding stuff to the
root without good cause would soon lead to overcrowding.

There is nothing structural about the repo that I'd want to key off of
as an alternative mechanism (because we might move packages around). We
could look inside each package.json we see to see if the name is
"liferay-frontend-projects", but the existence check implemented here
seems cheaper, so I went with that.