feat(js-toolkit): add a script to print out dependencies in topological order
My next step is going to be migrating the v3.x js-toolkit packages over
to the `@liferay` named scope on npm, so I need to do that in dependency
order. Now, I could decide this order by visual inspection, and make
that slightly easier with a command like:

    cat packages/*/package.json | jq .name,.dependencies

(Note that we only have to worry about "dependencies", not
"devDependencies", because they all got hoisted.)

However, I know we're going to need to do some other stuff that relies on
topological ordering so I made this small script that prints out the
packages in order like so:

    $ node scripts/topological.js
    no explicit target directory argument provided; assuming "packages"
    [
      "liferay-js-toolkit-core",
      "generator-liferay-js",
      "liferay-js-toolkit-scripts",
      "liferay-npm-bridge-generator",
      "liferay-npm-bundler"
    ]

ie. a valid ordering in which to publish the packages in the graph (or
in my case, a valid ordering in which to migrate them into the named
scope).

As per standard practice for command-line tools, the useful output (the
ordered list) goes to stdout:

    $ node scripts/topological.js 2> /dev/null
    [
      "liferay-js-toolkit-core",
      "generator-liferay-js",
      "liferay-js-toolkit-scripts",
      "liferay-npm-bridge-generator",
      "liferay-npm-bundler"
    ]

and the auxiliary output goes to stderr:

    $ node scripts/topological.js > /dev/null
    no explicit target directory argument provided; assuming "packages"

You can optionally pass a path to a directory to look for packages, as
you can see below, as well as noting that if you run it on a project
with a dependency cycle in it (eg. js-toolkit v2) it will print an error
because a topological ordering does not exist in the presence of cycles
(a topological ordering is defined as one in which every item is
guaranteed to appear before any and all things that depend on it):

    $ node scripts/topological.js ../../maintenance/projects/js-toolkit/packages
    Cycle detected in dependency graph

      babel-plugin-add-module-metadata

      babel-preset-liferay-standard

      liferay-npm-bundler-preset-standard

      liferay-npm-build-tools-common

      babel-plugin-add-module-metadata

    Error: Cycle detected in dependency graph

The package names involved in the cycle are printed as we unwind the
stack (ie. the opposite of the order in which they
were visited), which means you should interpret the above as
"babel-plugin-add-module-metadata is a dependency of
babel-preset-liferay-standard, which is a dependency of
liferay-npm-bundler-preset-standard... etc"

Note that the traversal algorithm is a slight variation on the one
presented in:

    https://en.wikipedia.org/wiki/Topological_sorting#Depth-first_search

That algorithm uses `unshift()` to prepend to the ordering array, and it
visits packages which depend on the node being visited. However, the
graph structure embodied in our package.json files doesn't explicitly
encode a list of "packages which depend on package X" but rather the
opposite ("packages which package X depends on").

So we use `push()` instead of `unshift()` and everything works out:
nothing gets `push()`-ed until the things it depends on have been
visited, either as a result of a recursive call to `visit()`, or because
they were already seen during a previous traversal (initiated from
inside our `forEach()` callback).