refactor(js-toolkit): rename generator-liferay-js to generator-js
Even though the Yeoman docs say that named scope packages are not
supported yet:

> The name should be generator-your-unique-name. This means it should
> begin with `generator-` and be followed by a unique (to npm)
> name. Scoped name support is coming in the future.

https://yeoman.io/generators/#content

They obviously are, as seen in working examples like these:

- https://www.npmjs.com/package/@dizmo/generator-module
- https://www.npmjs.com/package/@endemolshinegroup/generator-nodejs-ts
- https://www.npmjs.com/package/@pubcore/generator-rest

So, this commit prepares us to rename `generator-liferay-js` to
`@liferay/generator-js`. This is analogous to what we did with
`eslint-config-liferay`, which got renamed to `@liferay/eslint-config`.

Instead of:

    npm install -g yo generator-liferay-js
    yo liferay-js

Users will run:

    npm install -g yo @liferay/generator-js
    yo @liferay/js