refactor: document and simplify id check
Regexps are generally pretty hard to understand and especially so in
this case when we are relying on undocumented behavior on the part of
GitHub's Markdown renderer, so I added a comment to show what GitHub
does given the kind of input that we are supplying.

I made the following changes to the regexp:

- Dropped "m" flag, which does nothing in a regexp that uses neither `$`
  nor `^`.
- Used `\b` to detect word boundary before the `id` attribute;
  the old pattern (`[^i]*id`) would incorrectly match anything at all
  ending in `id`.
- Simplify tag name pattern from `[A-Za-z0-9]+` to `\w+`; I don't think
  numbers are legit in tag names.

Obviously, this test is not 100% rigorous (and parsing out HTML tags
with regexps is a fool's errand anyway), but it gets the job done and we
can tweak it as needed in the future.

One thing to note: give that GitHub is going to lowercase the IDs
anyway, we should probably write them in lowercase to begin with
(otherwise it gives the false impression that GitHub will preserve the
original case), but I figure that all this works and that could be left
for another commit.