refactor(changelog-generator): reformat usage information
As I am about to add several options here, I need to group these into
categories to make the quantity of information more manageable.

These are alphabetically sorted except for one anomaly: I made `--to`
follow `--from` because the options are so intimately related to each
other:

    --from=FROM                  Starting point (default: previous tag)
    --to=TO                      Ending point( default: "HEAD")

This is a bit subjective, and I kind of hate exceptions, but it was the
way it used to be before the grouping, so at least it remains mostly
consistent with its past ordering.