refactor(changelog-generator): "simplify" prerelease logic
"Simplify" is in quotes because any simplification is always debatable
and a bit subjective.

In this case, we're making it terser, which means the assignments can
be read in a more linear fashion, and we're getting rid of the unneeded
assignments to `preid`, which might be needless distracting.

As for the terseness, we have ternaries now; eg.

    prerelease = options.version === 'major' ? undefined : 0

I considered going further; eg.

    prerelease = options.version !== 'major' && 0;

but that seemed to be less readable. I also don't like the
polymorphic type (`boolean | number`) compared to the relatively
benign polymorphism of what we have now, which is `number | undefined`.