refactor(changelog-generator): streamline option handling
Because I am about to add yet another option, I want to pause to
refactor a bit because the "dumb and simple" approach that I had in
place initially was good at the time, but its verbosity is starting to
become bothersome as the number of options grows.

So, replace `option()` with a `matchOption()` function that enables us
to express things a little more tersely.

Note that I took the opportunity to fix a bug in the regular expressions
(they we expecting between 1 and 3 leading hyphens) and make them
looser: there's no reason not to accept all of:

- `--dry-run`
- `-dry-run`
- `dry-run`
- `--d`
- `-d`
- `d`

plus any of their case variations (eg. `--DRY-RUN` etc), so may as well
be lenient in what we accept.

And we may as well allow negation of any boolean option, not just
`--no-update-tags`. Now you can pass `--no-dry-run` (etc), which has the
effect of negating any previous `--dry-run` that has been passed.